refactor: split content models into dedicated module
